requirements.txt
pandas==2.2.2
numpy==1.26.4
plotly==5.23.0
streamlit==1.38.0
pytest==8.3.2
python-dateutil==2.9.0.post0